+++ GameEngineBase/src/Application/Window.h
@@ -13,8 +13,8 @@ namespace gebase {
 	{
 		uint width;
 		uint height;
-		uint halfWidth;
-		uint halfHeight;
+		long halfWidth;
+		long halfHeight;
 		bool fullscreen;
 		bool vsync;
 	};
@@ -50,8 +50,8 @@ namespace gebase {
 		inline uint getWidth() { return m_Properties.width; }
 		inline uint getHeight() { return m_Properties.height; }
 
-		inline uint getCenterX() { return m_Properties.halfWidth; }
-		inline uint getCenterY() { return m_Properties.halfHeight; }
+		inline long getCenterX() { return m_Properties.halfWidth; }
+		inline long getCenterY() { return m_Properties.halfHeight; }
 
 		void setVSync(bool vsync);
 		inline bool isVSync() { return m_VSync; }
